DHCP is working with my test setup
